fix(admin): 리뷰 반영 — range 검증(잘못된 range 400)·daysForRange 통합, salesSummary null-safe, 구독 1:1 주석, INFO 파일 ThresholdFilter

+++ src/test/kotlin/kr/ai/flori/admin/AdminUserIntegrationTest.kt
@@ -106,7 +106,7 @@ class AdminUserIntegrationTest {
         mockMvc
             .get("/admin/users/$uid") { header(HttpHeaders.AUTHORIZATION, "Bearer ${tokens.accessToken}") }
             .andExpect { status { isOk() } }
-            .andExpect { jsonPath("$.id") { value(uid.toInt()) } }
+            .andExpect { jsonPath("$.id") { value(uid) } }
             .andExpect { jsonPath("$.salesCount") { exists() } }
             .andExpect { jsonPath("$.verifications") { isArray() } }
             .andExpect { jsonPath("$.storeName") { exists() } }
